requirements.txt
google-cloud-bigquery==2.6.1
python-dotenv